requirements.txt
fastapi==0.122.0
uvicorn==0.38.0
supabase==2.24.0
google-generativeai==0.8.5
httpx==0.28.1
python-multipart==0.0.20
python-dotenv==1.2.1
